fix(telegram): extract bot info contract
